Lambda_deploy/app/db/database.py
# app/db/database.py
from pymongo import MongoClient
from dotenv import load_dotenv
import os

load_dotenv()

client = MongoClient(os.getenv("MONGODB_URI"))
db = client[os.getenv("DATABASE_NAME")]
users_collection = db[os.getenv("USERS_COLLECTION")]
