requirements.txt
fastapi
pydantic
mangum
uvicorn
passlib
python-jose
pydantic[email]
python-dotenv
pymongo
bcrypt